test: Unify karma files and update polyfills
