+++ packages/browser/test/integration/loader.html
@@ -4,9 +4,6 @@
 <head>
   <meta charset="utf-8">
   <title></title>
-  <script src="polyfills/es6-promise-4.2.4.js"></script>
-  <script src="polyfills/whatwg-fetch-2.0.4.js"></script>
-  <script src="common.js"></script>
   <script src="../../src/loader.js"></script>
   <script src="init.js"></script>
 </head>
